Refactor Terminator plugin structure and enhance command functionality
- Refactored code by separating concerns into distinct classes:
  - CommandClass: Handles command operations.
  - ListenerClass: Manages event listening.
  - ItemBuilderClass: Constructs and returns the Terminator item stack.
  - UtilsClass: Provides utilities, such as string color code conversion.
- Modified the main command to `/terminator getitem` for clarity.
- Enhanced documentation with JavaDoc comments for better maintainability and readability.

Questions for discussion:
1. Why is the Terminator item only usable by players?
2. What are the current project goals, and how can I contribute further to meet them?
